services/api_gateway/requirements.txt
Django==4.2
djangorestframework==3.14
pika==1.3.1  # RabbitMQ client
requests==2.28.1
redis==4.5.4
